Split sub-group tests into more files.
